Add TorchParametricProbabilisticForecastingModel
