+++ darts/models/tcn_model.py
@@ -15,7 +15,7 @@ from ..utils.data import ShiftedDataset
 from ..utils.likelihood_models import LikelihoodModel
 
 from ..logging import raise_if_not, get_logger
-from .torch_forecasting_model import TorchForecastingModel  # , _TimeSeriesShiftedDataset
+from .torch_forecasting_model import TorchParametricProbabilisticForecastingModel
 
 logger = get_logger(__name__)
 
@@ -121,8 +121,7 @@ class _TCNModule(nn.Module):
                  weight_norm: bool,
                  target_size: int,
                  target_length: int,
-                 dropout: float,
-                 likelihood: Optional[LikelihoodModel] = None):
+                 dropout: float):
 
         """ PyTorch module implementing a dilated TCN module used in `TCNModel`.
 
@@ -171,12 +170,9 @@ class _TCNModule(nn.Module):
         self.n_filters = num_filters
         self.kernel_size = kernel_size
         self.target_length = target_length
-        self.output_size = (
-            likelihood._num_parameters * target_size if likelihood is not None else target_size
-        )
+        self.output_size = target_size
         self.dilation_base = dilation_base
         self.dropout = nn.Dropout(p=dropout)
-        self.likelihood = likelihood
 
         # If num_layers is not passed, compute number of layers needed for full history coverage
         if num_layers is None and dilation_base > 1:
@@ -210,7 +206,7 @@ class _TCNModule(nn.Module):
         return x
 
 
-class TCNModel(TorchForecastingModel):
+class TCNModel(TorchParametricProbabilisticForecastingModel):
     @random_method
     def __init__(self,
                  input_chunk_length: int,
@@ -261,7 +257,7 @@ class TCNModel(TorchForecastingModel):
         kwargs['input_chunk_length'] = input_chunk_length
         kwargs['output_chunk_length'] = output_chunk_length
 
-        super().__init__(**kwargs)
+        super().__init__(likelihood=likelihood, **kwargs)
 
         self.input_chunk_length = input_chunk_length
         self.output_chunk_length = output_chunk_length
@@ -271,20 +267,21 @@ class TCNModel(TorchForecastingModel):
         self.dilation_base = dilation_base
         self.dropout = dropout
         self.weight_norm = weight_norm
-        self.likelihood = likelihood
 
     def _create_model(self, input_dim: int, output_dim: int) -> torch.nn.Module:
+        target_size = (
+            self.likelihood._num_parameters * output_dim if self.likelihood is not None else output_dim
+        )
         return _TCNModule(input_size=input_dim,
                           input_chunk_length=self.input_chunk_length,
-                          target_size=output_dim,
+                          target_size=target_size,
                           kernel_size=self.kernel_size,
                           num_filters=self.num_filters,
                           num_layers=self.num_layers,
                           dilation_base=self.dilation_base,
                           target_length=self.output_chunk_length,
                           dropout=self.dropout,
-                          weight_norm=self.weight_norm,
-                          likelihood=self.likelihood)
+                          weight_norm=self.weight_norm)
 
     def _build_train_dataset(self,
                              target: Sequence[TimeSeries],
@@ -294,21 +291,12 @@ class TCNModel(TorchForecastingModel):
                               length=self.input_chunk_length,
                               shift=self.output_chunk_length)
 
-    def _compute_loss(self, output, target):
-        if self.likelihood:
-            return self.likelihood._compute_loss(output, target)
-        else:
-            return super()._compute_loss(output, target)
-
     def _produce_predict_output(self, input):
         if self.likelihood:
             output = self.model(input)
             return self.likelihood._sample(output)
         else:
             return self.model(input)
-
-    def _is_probabilistic(self) -> bool:
-        return self.likelihood is not None
 
     @property
     def first_prediction_index(self) -> int:
